updated the api search function to have result render, search for album only
