+++ src/components/SearchFunction.js
@@ -2,6 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { TextField, Button, Menu, MenuItem } from '@mui/material';
 import SearchIcon from '@mui/icons-material/Search';
 import MenuIcon from '@mui/icons-material/Menu';
+import { Link } from 'react-router-dom';
+import SpotResults from './SpotResults';
 
 function SearchFunction() {
     const [searchTerm, setSearchTerm] = useState('');
@@ -12,6 +14,7 @@ function SearchFunction() {
         if (selectedOption === 'Search For Album' && searchTerm !== '') {
             // Perform search for album
             console.log('Performing search for album...');
+            localStorage.setItem("Search Term", searchTerm)
             searchSpotAlbum();
         } else if (selectedOption === 'Search For Sale' && searchTerm !== '') {
             // Perform search for sale
@@ -54,30 +57,32 @@ function SearchFunction() {
     };
 
     const searchSpotAlbum = async () => {
-        // Perform search logic using searchTerm
-        const url = `https://spotify23.p.rapidapi.com/search/?q=${searchTerm}&type=albums`;
+        // // Perform search logic using searchTerm
+        // const url = `https://spotify23.p.rapidapi.com/search/?q=${encodeURIComponent(
+        //     searchTerm
+        // )}&type=albums`;
 
-        const options = {
-            method: 'GET',
-            headers: {
-                'X-RapidAPI-Key': process.env.REACT_APP_SPOTIFY_API_KEY,
-                'X-RapidAPI-Host': 'spotify23.p.rapidapi.com',
-            },
-        };
+        // const options = {
+        //     method: 'GET',
+        //     headers: {
+        //         'X-RapidAPI-Key': process.env.REACT_APP_SPOTIFY_API_KEY,
+        //         'X-RapidAPI-Host': 'spotify23.p.rapidapi.com',
+        //     },
+        // };
 
-        try {
-            const response = await fetch(url, options);
+        // try {
+        //     const response = await fetch(url, options);
 
-            if (response.ok) {
-                const result = await response.json();
-                console.log(result);
-            } else {
-                console.error('Error occurred while searching');
-            }
-        } catch (error) {
-            console.error('Error occurred while searching', error);
-        }
-
+        //     if (response.ok) {
+        //         const result = await response.json();
+        //         console.log(result);
+        //     } else {
+        //         console.error('Error occurred while searching');
+        //     }
+        // } catch (error) {
+        //     console.error('Error occurred while searching', error);
+        // }
+        <SpotResults />
     };
 
     const searchDiscogsAlbum = async () => {
@@ -104,32 +109,6 @@ function SearchFunction() {
         }
     };
 
-    const handleSearch = (selectedOption) => {
-        switch (selectedOption) {
-            case 'Search For Album':
-                // Perform search for album
-                console.log('Performing search for album...');
-                searchSpotAlbum(localStorage.getItem("Search Term:"));
-                break;
-            case 'Search For Sale':
-                // Perform search for sale
-                console.log('Performing search for sale...');
-                searchDiscogsAlbum(localStorage.getItem("Search Term:"))
-                break;
-            case 'Get All Users':
-                // Get all users
-                console.log('Getting all users...');
-                break;
-            case 'Get Single User':
-                // Get single user
-                console.log('Getting single user...');
-                break;
-            default:
-                console.log('Invalid search option');
-                break;
-        }
-    };
-
     return (
         <form onSubmit={handleSubmit} style={{ display: 'flex', alignItems: 'center' }}>
             <TextField
@@ -150,7 +129,7 @@ function SearchFunction() {
                 }}
             />
             <Menu anchorEl={anchorEl} open={Boolean(anchorEl)} onClose={handleMenuClose}>
-                <MenuItem onClick={() => handleOptionSelect('Search For Album')}>Search For Album</MenuItem>
+                <MenuItem onClick={() => handleOptionSelect('Search For Album')} component={Link} to="/search/spot">Search For Album</MenuItem>
                 <MenuItem onClick={() => handleOptionSelect('Search For Sale')}>Search For Sale</MenuItem>
                 <MenuItem onClick={() => handleOptionSelect('Get All Users')}>Get All Users</MenuItem>
                 <MenuItem onClick={() => handleOptionSelect('Get Single User')}>Get Single User</MenuItem>
